account section refactored
