+++ app/src/main/java/org/mifos/selfserviceapp/api/services/LoanAccountsListService.java
@@ -1,7 +1,7 @@
 package org.mifos.selfserviceapp.api.services;
 
 import org.mifos.selfserviceapp.api.ApiEndPoints;
-import org.mifos.selfserviceapp.data.accounts.LoanAccount;
+import org.mifos.selfserviceapp.data.accounts.LoanAccountsListResponse;
 
 import retrofit2.Call;
 import retrofit2.http.GET;
@@ -11,8 +11,7 @@ import retrofit2.http.Path;
  * @author Vishwajeet
  * @since 23/6/16.
  */
-
 public interface LoanAccountsListService {
     @GET(ApiEndPoints.CLIENTS + "/{clientId}/accounts?fields=loanAccounts")
-    Call<LoanAccount> getLoanAccountsList(@Path("clientId") int clientId);
+    Call<LoanAccountsListResponse> getLoanAccountsList(@Path("clientId") int clientId);
 }
