+++ app/src/main/java/org/mifos/selfserviceapp/data/accounts/LoanAccount.java
@@ -1,8 +1,5 @@
 package org.mifos.selfserviceapp.data.accounts;
 
-import java.util.ArrayList;
-import java.util.List;
-
 /**
  * @author Vishwajeet
  * @since 22/06/16.
@@ -11,15 +8,6 @@ public class LoanAccount {
     private int id;
     private String accountNo;
     private String productName;
-    private List<LoanAccount> loanAccounts = new ArrayList<LoanAccount>();
-
-    public List<LoanAccount> getLoanAccounts() {
-        return loanAccounts;
-    }
-
-    public void setLoanAccounts(List<LoanAccount> loanAccounts) {
-        this.loanAccounts = loanAccounts;
-    }
 
     public String getAccountNo() {
         return accountNo;
