+++ app/src/main/java/org/mifos/selfserviceapp/data/accounts/LoanAccountsListResponse.java
@@ -1,0 +1,20 @@
+package org.mifos.selfserviceapp.data.accounts;
+
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * @author Vishwajeet
+ * @since 13/08/16
+ */
+public class LoanAccountsListResponse {
+    private List<LoanAccount> loanAccounts = new ArrayList<>();
+
+    public List<LoanAccount> getLoanAccounts() {
+        return loanAccounts;
+    }
+
+    public void setAccounts(List<LoanAccount> loanAccounts) {
+        this.loanAccounts = loanAccounts;
+    }
+}
