+++ app/src/main/java/org/mifos/selfserviceapp/data/accounts/SavingAccount.java
@@ -1,8 +1,5 @@
 package org.mifos.selfserviceapp.data.accounts;
 
-import java.util.ArrayList;
-import java.util.List;
-
 /**
  * @author Vishwajeet
  * @since 22/06/16
@@ -12,15 +9,6 @@ public class SavingAccount {
     private int id;
     private String accountNo;
     private String productName;
-    private List<SavingAccount> savingsAccounts = new ArrayList<SavingAccount>();
-
-    public List<SavingAccount> getSavingsAccounts() {
-        return savingsAccounts;
-    }
-
-    public void setSavingsAccounts(List<SavingAccount> savingsAccounts) {
-        this.savingsAccounts = savingsAccounts;
-    }
 
     public String getAccountNo() {
         return accountNo;
